cesem/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap3==2022.1
Django==4.2
django-crispy-forms==2.0
django-filter==23.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
xlrd==2.0.1
